+++ CHANGES.md
@@ -36,6 +36,7 @@ Changes in 11.3
 - Use DBUpsert() & check for required columns on INSERT in Medical.inc.php
 - Fix use DBUnescapeString() to unescape string for both PostgreSQL & MySQL in Config.fnc.php
 - Fix sorting list when first value is null in ListOutput.fnc.php
+- JS fix rt2colorBox children height calculation in warehouse.js
 
 Changes in 11.2.4
 -----------------